#1860 Remove impl pattern for network components
