+++ dynawo/sources/Models/CPP/ModelNetwork/DYNModelGenerator.h
@@ -21,7 +21,7 @@
 #define MODELS_CPP_MODELNETWORK_DYNMODELGENERATOR_H_
 
 #include <boost/shared_ptr.hpp>
-#include "DYNNetworkComponentImpl.h"
+#include "DYNNetworkComponent.h"
 
 namespace DYN {
 class ModelBus;
@@ -30,7 +30,7 @@ class GeneratorInterface;
 /**
  * class ModelGenerator
  */
-class ModelGenerator : public NetworkComponent::Impl {
+class ModelGenerator : public NetworkComponent {
  public:
   /**
    * @brief default constructor
@@ -87,17 +87,17 @@ class ModelGenerator : public NetworkComponent::Impl {
   void evalDerivativesPrim() { /* not needed */ }
 
   /**
-   * @copydoc NetworkComponent::Impl::evalF()
+   * @copydoc NetworkComponent::evalF()
    */
   void evalF(propertyF_t type);
 
   /**
-   * @copydoc NetworkComponent::Impl::evalJt(SparseMatrix& jt, const double& cj, const int& rowOffset)
+   * @copydoc NetworkComponent::evalJt(SparseMatrix& jt, const double& cj, const int& rowOffset)
    */
   void evalJt(SparseMatrix& jt, const double& cj, const int& rowOffset);
 
   /**
-   * @copydoc NetworkComponent::Impl::evalJtPrim(SparseMatrix& jt, const int& rowOffset)
+   * @copydoc NetworkComponent::evalJtPrim(SparseMatrix& jt, const int& rowOffset)
    */
   void evalJtPrim(SparseMatrix& jt, const int& rowOffset);
 
@@ -133,7 +133,7 @@ class ModelGenerator : public NetworkComponent::Impl {
   void defineElements(std::vector<Element>& elements, std::map<std::string, int>& mapElement);
 
   /**
-   * @copydoc NetworkComponent::Impl::evalZ()
+   * @copydoc NetworkComponent::evalZ()
    */
   NetworkComponent::StateChange_t evalZ(const double& t);
 
@@ -207,12 +207,12 @@ class ModelGenerator : public NetworkComponent::Impl {
   void init(int & yNum);
 
   /**
-   * @copydoc NetworkComponent::Impl::getY0()
+   * @copydoc NetworkComponent::getY0()
    */
   void getY0();
 
   /**
-   * @copydoc NetworkComponent::Impl::setSubModelParameters(const boost::unordered_map<std::string, ParameterModeler>& params)
+   * @copydoc NetworkComponent::setSubModelParameters(const boost::unordered_map<std::string, ParameterModeler>& params)
    */
   void setSubModelParameters(const boost::unordered_map<std::string, ParameterModeler>& params);
