+++ dynawo/sources/Models/CPP/ModelNetwork/DYNModelThreeWindingsTransformer.h
@@ -21,7 +21,7 @@
 #define MODELS_CPP_MODELNETWORK_DYNMODELTHREEWINDINGSTRANSFORMER_H_
 
 #include <boost/shared_ptr.hpp>
-#include "DYNNetworkComponentImpl.h"
+#include "DYNNetworkComponent.h"
 
 namespace DYN {
 class ModelBus;
@@ -30,7 +30,7 @@ class ThreeWTransformerInterface;
 /**
  * class Three Windings Transformer Model
  */
-class ModelThreeWindingsTransformer : public NetworkComponent::Impl {
+class ModelThreeWindingsTransformer : public NetworkComponent {
  public:
   /**
    * @brief default constructor
@@ -118,37 +118,37 @@ class ModelThreeWindingsTransformer : public NetworkComponent::Impl {
   void evalDerivatives(const double /*cj*/) { /* not needed */ }
 
   /**
-   * @copydoc NetworkComponent::Impl::evalDerivativesPrim()
+   * @copydoc NetworkComponent::evalDerivativesPrim()
    */
   void evalDerivativesPrim() { /* not needed */ }
 
   /**
-   * @copydoc NetworkComponent::Impl::evalF()
+   * @copydoc NetworkComponent::evalF()
    */
   void evalF(propertyF_t type);
 
   /**
-   * @copydoc NetworkComponent::Impl::evalJt(SparseMatrix& jt, const double& cj, const int& rowOffset)
+   * @copydoc NetworkComponent::evalJt(SparseMatrix& jt, const double& cj, const int& rowOffset)
    */
   void evalJt(SparseMatrix& jt, const double& cj, const int& rowOffset);
 
   /**
-   * @copydoc NetworkComponent::Impl::evalJtPrim(SparseMatrix& jt, const int& rowOffset)
+   * @copydoc NetworkComponent::evalJtPrim(SparseMatrix& jt, const int& rowOffset)
    */
   void evalJtPrim(SparseMatrix& jt, const int& rowOffset);
 
   /**
-   * @copydoc NetworkComponent::Impl::defineElements()
+   * @copydoc NetworkComponent::defineElements()
    */
   void defineElements(std::vector<Element>& elements, std::map<std::string, int>& mapElement);
 
   /**
-   * @copydoc NetworkComponent::Impl::evalZ()
+   * @copydoc NetworkComponent::evalZ()
    */
   NetworkComponent::StateChange_t evalZ(const double& t);
 
   /**
-   * @copydoc NetworkComponent::Impl::evalG()
+   * @copydoc NetworkComponent::evalG()
    */
   void evalG(const double& t);
 
@@ -217,12 +217,12 @@ class ModelThreeWindingsTransformer : public NetworkComponent::Impl {
   void init(int& yNum);
 
   /**
-   * @copydoc NetworkComponent::Impl::getY0()
+   * @copydoc NetworkComponent::getY0()
    */
   void getY0() { /* not needed */ }
 
   /**
-   * @copydoc NetworkComponent::Impl::setSubModelParameters(const boost::unordered_map<std::string, ParameterModeler>& params)
+   * @copydoc NetworkComponent::setSubModelParameters(const boost::unordered_map<std::string, ParameterModeler>& params)
    */
   void setSubModelParameters(const boost::unordered_map<std::string, ParameterModeler>& params);
