+++ dynawo/sources/Models/CPP/ModelNetwork/DYNModelTwoWindingsTransformer.cpp
@@ -63,7 +63,7 @@ using boost::shared_ptr;
 namespace DYN {
 
 ModelTwoWindingsTransformer::ModelTwoWindingsTransformer(const shared_ptr<TwoWTransformerInterface>& tfo) :
-Impl(tfo->getID()),
+NetworkComponent(tfo->getID()),
 ir1_dUr1_(0.),
 ir1_dUi1_(0.),
 ir1_dUr2_(0.),
