+++ dynawo/sources/Models/CPP/ModelNetwork/DYNNetworkComponent.hpp
@@ -11,42 +11,40 @@
 //
 
 /**
- * @file  DYNNetworkComponentImpl.hpp
+ * @file  DYNNetworkComponent.hpp
  *
  * @brief Implementation of template methods
  *
  */
 
-#ifndef MODELS_CPP_MODELNETWORK_DYNNETWORKCOMPONENTIMPL_HPP_
-#define MODELS_CPP_MODELNETWORK_DYNNETWORKCOMPONENTIMPL_HPP_
-
-#include "DYNParameterModeler.h"
+#ifndef MODELS_CPP_MODELNETWORK_DYNNETWORKCOMPONENT_HPP_
+#define MODELS_CPP_MODELNETWORK_DYNNETWORKCOMPONENT_HPP_
 
 namespace DYN {
 
 template <>
 double
-NetworkComponent::Impl::getParameterDynamicNoThrow(const boost::unordered_map<std::string, ParameterModeler>&,
+NetworkComponent::getParameterDynamicNoThrow(const boost::unordered_map<std::string, ParameterModeler>&,
                                                    const std::string&, bool&, const std::vector<std::string>&) const;
 
 template <>
 int
-NetworkComponent::Impl::getParameterDynamicNoThrow(const boost::unordered_map<std::string, ParameterModeler>&,
+NetworkComponent::getParameterDynamicNoThrow(const boost::unordered_map<std::string, ParameterModeler>&,
                                                    const std::string&, bool&, const std::vector<std::string>&) const;
 
 template <>
 bool
-NetworkComponent::Impl::getParameterDynamicNoThrow(const boost::unordered_map<std::string, ParameterModeler>&,
+NetworkComponent::getParameterDynamicNoThrow(const boost::unordered_map<std::string, ParameterModeler>&,
                                                    const std::string&, bool&, const std::vector<std::string>&) const;
 
 template <>
 std::string
-NetworkComponent::Impl::getParameterDynamicNoThrow(const boost::unordered_map<std::string, ParameterModeler>&,
+NetworkComponent::getParameterDynamicNoThrow(const boost::unordered_map<std::string, ParameterModeler>&,
                                                    const std::string&, bool&, const std::vector<std::string>&) const;
 
 template<typename T>
 T
-inline NetworkComponent::Impl::getParameterDynamic(const boost::unordered_map<std::string, ParameterModeler>& params,
+inline NetworkComponent::getParameterDynamic(const boost::unordered_map<std::string, ParameterModeler>& params,
     const std::string& id, const std::vector<std::string>& ids) const {
   bool foundParam = false;
   T value = getParameterDynamicNoThrow<T>(params, id, foundParam, ids);
@@ -58,7 +56,7 @@ inline NetworkComponent::Impl::getParameterDynamic(const boost::unordered_map<st
 
 template<typename T>
 void
-inline NetworkComponent::Impl::findParameterDynamicNoThrow(const boost::unordered_map<std::string, ParameterModeler>& params,
+inline NetworkComponent::findParameterDynamicNoThrow(const boost::unordered_map<std::string, ParameterModeler>& params,
     const std::string& id, bool& foundParam, const std::vector<std::string>& ids, T& value) const {
   foundParam = false;
   if (ids.empty()) {
@@ -87,4 +85,4 @@ inline NetworkComponent::Impl::findParameterDynamicNoThrow(const boost::unordere
 
 }  // namespace DYN
 
-#endif  // MODELS_CPP_MODELNETWORK_DYNNETWORKCOMPONENTIMPL_HPP_
+#endif  // MODELS_CPP_MODELNETWORK_DYNNETWORKCOMPONENT_HPP_
